% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_spline.R
\name{fit_spline}
\alias{fit_spline}
\title{Fit mixed effects spline model}
\usage{
fit_spline(i, t, f, A)
}
\arguments{
\item{i}{factor denoting what individual the sample is from}

\item{t}{numeric denoting time of sampling}

\item{f}{numeric feature value}

\item{A}{design matrix for spline basis}
}
\value{
list containing mod (fittet model object), pval (p-value for test of significant spline) and coef (coefficients from fitted model)
}
\description{
Fits a spline model to longitudinal data with random effects for individual baseline levels
}
