% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laap.R
\name{laap}
\alias{laap}
\title{Fit mixed effects spline models for all analytes}
\usage{
laap(i, t, L, k)
}
\arguments{
\item{i}{factor denoting what individual the sample is from}

\item{t}{numeric denoting time of sampling}

\item{L}{numeric matrix of feature values / analyte levels}

\item{k}{number of knots in spline basis}
}
\value{
list containing pval (p-values for test of significant spline) and coef (coefficients from fitted model)
}
\description{
Fits a spline model to longitudinal data with random effects for individual baseline levels for all analytes in dataframe
}
