% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_test_data.R
\name{sim_meta_data}
\alias{sim_meta_data}
\title{Simulation of sample meta data}
\usage{
sim_meta_data(n = 20, m = 30, r1 = 0, r2 = 40 * 7)
}
\arguments{
\item{n}{integer number of individuals}

\item{m}{integer number of samples per individual}

\item{r1}{integer(days) start of possible sampling period}

\item{r2}{integer(days) end of possible sampling period}
}
\value{
Data-frame of meta (individual and timepoint) data for each sample
}
\description{
Simulates m random sample timepoints for n individuals
}
