% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_test_data.R
\name{sim_single_analyte}
\alias{sim_single_analyte}
\title{Simulate measurements of single analyte}
\usage{
sim_single_analyte(i, t, v1 = 1, v2 = 1, t1 = 20 * 7, t2 = 25 * 7,
  h = 6)
}
\arguments{
\item{i}{vector naming the individual to which the sample belongs}

\item{t}{vector timepoint sample was taken}

\item{v1}{standard deviation for random effect in individual baseline level}

\item{v2}{standard deviation for measurement noise}

\item{t1}{start timepoint for increase in analyte level}

\item{t2}{timepoint for return to baseline individual level}

\item{h}{max level for analyte (reached just prior to t2)}
}
\value{
Analyte levels (same order as input order)
}
\description{
Simulate measurements of single analyte based on individual and timepoint for the sample
}
