% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_test_data.R
\name{sim_test_data}
\alias{sim_test_data}
\title{Simulate test dataset for longitudinal analysis of analyte patterns}
\usage{
sim_test_data(n = 20, m = 30, r1 = 0, r2 = 40 * 7, v1 = 1, v2 = 1,
  t1 = rep(20 * 7, 500), t2 = rep(25 * 7, 500), h = rep(c(0, 6), each =
  250))
}
\arguments{
\item{n}{integer number of individuals}

\item{m}{integer number of samples per individual}

\item{r1}{integer(days) start of possible sampling period}

\item{r2}{integer(days) end of possible sampling period}

\item{v1}{standard deviation for random effect in individual baseline level}

\item{v2}{standard deviation for measurement noise}

\item{t1}{start timepoint for increase in analyte level}

\item{t2}{timepoint for return to baseline individual level}

\item{h}{max level for analyte (reached just prior to t2)}
}
\value{
List with meta data (individual and timepoint) for each sample and analytes (analyte levels)
}
\description{
Simulate random sampling time and analyte level for set of analytes
}
